solution folder created for zoom reports solution
