+++ Solutions/ZoomReports/Data/Solution_ZoomReports.json
@@ -1,0 +1,18 @@
+{
+  "Name": "ZoomReports",
+  "Author": "Microsoft - support@microsoft.com",
+  "Logo": "<img src=\"https://raw.githubusercontent.com/Azure/Azure-Sentinel/master/Logos/Azure_Sentinel.svg\" width=\"75px\" height=\"75px\">",
+  "Description": "The [Zoom](https://zoom.us/) Reports solution enables you to ingest Zoom Reports' events into Microsoft Sentinel through the [Zoom Report REST API](https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/reportSignInSignOutActivities)\r\n  \r\n  **Underlying Microsoft Technologies used:**\r\n\n  This solution takes a dependency on the following technologies, and some of these dependencies either may be in [Preview](https://azure.microsoft.com/en-us/support/legal/preview-supplemental-terms/) state or might result in additional ingestion or operational costs:\r\n\n  a.\t [Azure Monitor HTTP Data Collector API](https://docs.microsoft.com/en-us/azure/azure-monitor/logs/data-collector-api)\r\n\n  b.\t [Azure Functions](https://azure.microsoft.com/en-us/services/functions/#overview)",
+  "Data Connectors": [
+    "Solutions/ZoomReports/Data Connectors/ZoomReports/azuredeploy_Connector_ZoomAPI_AzureFunction.json"
+  ],
+  "Parsers": [
+    "Solutions/ZoomReports/Parsers/Zoom.txt"
+  ],
+  
+  "BasePath": "C:\\GitHub\\Azure-Sentinel",
+  "Version": "2.0.0",
+  "Metadata": "SolutionMetadata.json",
+  "TemplateSpec": true,
+  "Is1PConnector": false
+}
