+++ Solutions/ZoomReports/Package/createUiDefinition.json
@@ -1,0 +1,92 @@
+{
+    "$schema": "https://schema.management.azure.com/schemas/0.1.2-preview/CreateUIDefinition.MultiVm.json#",
+    "handler": "Microsoft.Azure.CreateUIDef",
+    "version": "0.1.2-preview",
+    "parameters": {
+      "config": {
+        "isWizard": false,
+        "basics": {
+          "description": "**Note:** _There may be [known issues](https://aka.ms/sentinelsolutionsknownissues) pertaining to this Solution, please refer to them before installing._\n\nThe [Zoom](https://zoom.us/) Reports solution enables you to ingest Zoom Reports' events into Microsoft Sentinel through the [Zoom Report REST API](https://marketplace.zoom.us/docs/api-reference/zoom-api/methods/#operation/reportSignInSignOutActivities)\r\n  \r\n  **Underlying Microsoft Technologies used:**\r\n\n  This solution takes a dependency on the following technologies, and some of these dependencies either may be in [Preview](https://azure.microsoft.com/en-us/support/legal/preview-supplemental-terms/) state or might result in additional ingestion or operational costs:\r\n\n  a.\t [Azure Monitor HTTP Data Collector API](https://docs.microsoft.com/en-us/azure/azure-monitor/logs/data-collector-api)\r\n\n  b.\t [Azure Functions](https://azure.microsoft.com/en-us/services/functions/#overview)  \n\n**Data Connectors:** 1, **Parsers:** 1\n\n[Learn more about Microsoft Sentinel](https://aka.ms/azuresentinel) | [Learn more about Solutions](https://aka.ms/azuresentinelsolutionsdoc)",
+          "subscription": {
+            "resourceProviders": [
+              "Microsoft.OperationsManagement/solutions",
+              "Microsoft.OperationalInsights/workspaces/providers/alertRules",
+              "Microsoft.Insights/workbooks",
+              "Microsoft.Logic/workflows"
+            ]
+          },
+          "location": {
+            "metadata": {
+              "hidden": "Hiding location, we get it from the log analytics workspace"
+            },
+            "visible": false
+          },
+          "resourceGroup": {
+            "allowExisting": true
+          }
+        }
+      },
+      "basics": [
+        {
+          "name": "getLAWorkspace",
+          "type": "Microsoft.Solutions.ArmApiControl",
+          "toolTip": "This filters by workspaces that exist in the Resource Group selected",
+          "condition": "[greater(length(resourceGroup().name),0)]",
+          "request": {
+            "method": "GET",
+            "path": "[concat(subscription().id,'/providers/Microsoft.OperationalInsights/workspaces?api-version=2020-08-01')]"
+          }
+        },
+        {
+          "name": "workspace",
+          "type": "Microsoft.Common.DropDown",
+          "label": "Workspace",
+          "placeholder": "Select a workspace",
+          "toolTip": "This dropdown will list only workspace that exists in the Resource Group selected",
+          "constraints": {
+            "allowedValues": "[map(filter(basics('getLAWorkspace').value, (filter) => contains(toLower(filter.id), toLower(resourceGroup().name))), (item) => parse(concat('{\"label\":\"', item.name, '\",\"value\":\"', item.name, '\"}')))]",
+            "required": true
+          },
+          "visible": true
+        }
+      ],
+      "steps": [
+        {
+          "name": "dataconnectors",
+          "label": "Data Connectors",
+          "bladeTitle": "Data Connectors",
+          "elements": [
+            {
+              "name": "dataconnectors1-text",
+              "type": "Microsoft.Common.TextBlock",
+              "options": {
+                "text": "This solution installs the data connector for ingesting Zoom Reports events into Microsoft Sentinel using the Zoom Reports REST API. The events captured provide the ability to examine potential security risks, analyze your team's use of collaboration, diagnose configuration problems and more. After installing the solution, configure and enable this data connector by following guidance in Manage solution view."
+              }
+            },
+            {
+              "name": "dataconnectors-parser-text",
+              "type": "Microsoft.Common.TextBlock",
+              "options": {
+                "text": "The solution installs a parser that transforms ingested data. The transformed logs can be accessed using the Zoom Kusto Function alias."
+              }
+            },
+            {
+              "name": "dataconnectors-link1",
+              "type": "Microsoft.Common.TextBlock",
+              "options": {
+                "link": {
+                  "label": "Learn more about connecting data sources",
+                  "uri": "https://docs.microsoft.com/azure/sentinel/connect-data-sources"
+                }
+              }
+            }
+          ]
+        }
+      ],
+      "outputs": {
+        "workspace-location": "[first(map(filter(basics('getLAWorkspace').value, (filter) => and(contains(toLower(filter.id), toLower(resourceGroup().name)),equals(filter.name,basics('workspace')))), (item) => item.location))]",
+        "location": "[location()]",
+        "workspace": "[basics('workspace')]"
+      }
+    }
+  }
