+++ Solutions/ZoomReports/SolutionMetadata.json
@@ -1,0 +1,16 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-zoomreports",
+	"firstPublishDate": "2022-05-23",
+	"providers": ["Zoom"],
+	"categories": {
+		"domains" : ["Application"],
+		"verticals": []
+	},
+	"support": {
+		"tier": "Microsoft",
+		"name": "Microsoft Corporation",
+		"email": "support@microsoft.com",
+		"link": "https://support.microsoft.com/"
+	}
+}
